# expression: Implement `VecEvalBool` to prepare for vectorized selection operator