@qw4990 2019-09-16
/run-sqllogic-test-1